Placed Cmds in Folder Structure
