+++ RapidReact2606/src/main/java/frc/robot/RobotContainer.java
@@ -13,12 +13,12 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.XboxController.Axis;
 import frc.robot.Constants.IOConstants;
-import frc.robot.commands.AutoDriveBack;
-import frc.robot.commands.OneBallAuto;
-import frc.robot.commands.SimpleIndexerOn;
-import frc.robot.commands.SimpleIntakeOn;
-import frc.robot.commands.SimpleIntakeOnVar;
-import frc.robot.commands.SimpleShooterOn;
+import frc.robot.commands.Auto.AutoDriveBack;
+import frc.robot.commands.Auto.OneBallAuto;
+import frc.robot.commands.Indexer.SimpleIndexerOn;
+import frc.robot.commands.Intake.SimpleIntakeOn;
+import frc.robot.commands.Intake.SimpleIntakeOnVar;
+import frc.robot.commands.Shooter.SimpleShooterOn;
 // import frc.robot.commands.ExampleCommand;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.IndexerSubsystem;
