+++ RapidReact2606/src/main/java/frc/robot/commands/Auto/AutoDriveBack.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.Auto;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DriveSubsystem;
