+++ RapidReact2606/src/main/java/frc/robot/commands/Auto/OneBallAuto.java
@@ -1,7 +1,10 @@
-package frc.robot.commands;
+package frc.robot.commands.Auto;
 
 // import edu.wpi.first.wpilibj2.command.CommandGroupBase;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.commands.Indexer.SimpleIndexerOn;
+import frc.robot.commands.Shooter.SetShooterMode;
+import frc.robot.commands.Shooter.isShooterUpToSpeed;
 import frc.robot.subsystems.IndexerSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
 import frc.robot.subsystems.IndexerSubsystem.MODE;
