+++ RapidReact2606/src/main/java/frc/robot/commands/Shooter/SimpleShooterOn.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.Shooter;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ShooterSubsystem;
