+++ RapidReact2606/src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -13,7 +13,7 @@ import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants.ShooterConstants;
-import frc.robot.commands.isShooterUpToSpeed;
+import frc.robot.commands.Shooter.isShooterUpToSpeed;
 
 public class ShooterSubsystem extends SubsystemBase {
   CANSparkMax shooter = new CANSparkMax(ShooterConstants.shooterCan, MotorType.kBrushless);
